Rename cli helper
