+++ lib/helpers/logger.js
@@ -1,9 +1,9 @@
 const bunyan = require('bunyan');
-const CliHelper = require('./cli');
+const PrettyStdout = require('./logger-pretty-stdout');
 const traverse = require('traverse');
 
-const cliHelper = new CliHelper();
-cliHelper.pipe(process.stdout);
+const prettyStdOut = new PrettyStdout();
+prettyStdOut.pipe(process.stdout);
 
 function configSerializer(config) {
   const redactedFields = ['token', 'githubAppKey'];
@@ -29,7 +29,7 @@ const logger = bunyan.createLogger({
       name: 'stdout',
       level: process.env.LOG_LEVEL || 'info',
       type: 'raw',
-      stream: cliHelper,
+      stream: prettyStdOut,
     },
   ],
 });
